[FE][refactor] : Milestonelist component분리 #21
MilestoneListPage에서 마일스톤리스트 부분을 component로 분리
